Added saving and loading to BfsResult
